// frontend/src/features/photos/types.ts
export interface Photo {
	alt: string;
	avg_color: string;
	height: number;
	id: number;
	photographer: string;
	photographer_id: number;
	photographer_url: string;
	src_landscape: string;
	src_large: string;
	src_large2x: string;
	src_medium: string;
	src_original: string;
	src_portrait: string;
	src_small: string;
	src_tiny: string;
	url: string;
	width: number;
	is_liked: boolean;
}